Duplicate system sheet to all characters instead of saving just the basic
It was decided to follow this direction even though it will increase the amount of data saved in the database (and, furthermore, increasing the overhead of reading data from the database). This solution was adopted because it is the simplest one considering maintainability, and also it will save a whole lot of processing every time a sheet is read from the database. Without this, every time the sheet applies the relationships among the attributes groups it would have to either replicate the values from the game system sheet of to modify itself to transform arrays into hashes, speeding up the performance on every find that would be made to every single attribute (every character attribute would have to be matched with the one on the system sheet).

+++ app/models/rpg/attributes_group.rb
@@ -39,7 +39,7 @@ module RPG
         @used_points ||= begin
           return 0 if character_attributes.blank?
 
-          character_attributes.reduce(0) {|sum, attribute| sum += attribute.points.to_f }
+          character_attributes.reduce(0) {|sum, attribute| sum += attribute.points.to_i }
         end
       end
